// src/lib/clients/jikan/generated/models/anime_meta.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { anime_images } from './anime_images';

export type anime_meta = {
    /**
     * MyAnimeList ID
     */
    mal_id?: number;
    /**
     * MyAnimeList URL
     */
    url?: string;
    images?: anime_images;
    /**
     * Entry title
     */
    title?: string;
};