// src/lib/clients/jikan/generated/models/user_by_id.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * User Meta By ID
 */
export type user_by_id = {
    /**
     * MyAnimeList URL
     */
    url?: string;
    /**
     * MyAnimeList Username
     */
    username?: string;
};